+++ pypeit/tests/test_chk_calibs.py
@@ -27,9 +27,15 @@ def test_chk_calibs_not():
 
 @dev_suite_required
 def test_chk_calibs_deimos():
-    droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/keck_deimos/*/')
+    # 830G
+    droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/keck_deimos/830G_M_8600/')
     pargs = chk_calibs.parser([droot, '-s', 'keck_deimos'])
     answers, _ = chk_calibs.main(pargs)
+    assert np.all(answers['pass'][:-1]), 'One or more failures!'
 
+    # 600ZD
+    droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/keck_deimos/600ZD_M_7500/')
+    pargs = chk_calibs.parser([droot, '-s', 'keck_deimos'])
+    answers, _ = chk_calibs.main(pargs)
     assert np.all(answers['pass'][:-1]), 'One or more failures!'