+++ CHANGES.rst
@@ -3,7 +3,7 @@
 --------
  - Fixed a bug fixes a bug in full_template wavelength reidentification for situations where extreme
    wavelength coverage slits results in reidentification with a purely zero-padded array.
-
+ - (Hotfix) Deal with chk_calibs test
 
 1.0.6 (22 Jul 2020)
 -------------------